code cleanup and edits to readme.md
